FirstHandIn/requirements.txt
pillow
stepic